Implemented decompression

+++ rle.h
@@ -1,6 +1,7 @@
 #ifndef RLE_H
 #define RLE_H
 #include <stdio.h>
+#include <stdint.h>
 
 #define RLE_DEFAULT_BUFFER_SIZE 8192
 #define RLE_MIN_RUN_LENGTH 3      // minimum length for an encoding run
@@ -9,9 +10,9 @@
 #define RLE_SIGNATURE "RLE1"      // format signature
 
 typedef struct {
-    unsigned long buffer_size;
-    unsigned long total_bytes_processed;
-    unsigned long total_bytes_output;  
+    uint32_t buffer_size;
+    uint32_t total_bytes_processed;
+    uint32_t total_bytes_output;  
     unsigned char* buffer;
     FILE* input;
     FILE* output;
